Fix runtime error when using gtest_filter on specific tests.
Kokkos threw a runtime error complaining about deallocation after a
Kokkos::finalize call. This occurred because the master elements created
for specific tests weren't deallocated prior to exit. These elements
were deallocated by `Realm::~Realm()` when the full-suite was run and we
were lucky in the ordering of tests such that the tests with realm
destructors run at the end cleaning up the memory.

Fixes #196.